+++ presto-orc/src/main/java/com/facebook/presto/orc/stream/ByteStream.java
@@ -98,7 +98,7 @@ public class ByteStream
     }
 
     @Override
-    public void skip(int items)
+    public void skip(long items)
             throws IOException
     {
         while (items > 0) {
